requirements.txt
APScheduler==3.9.1
asgiref==3.5.2
autopep8==2.0.0
base58==2.1.1
bitcoin-utils-fork-minimal==0.4.11.6
block-io==2.0.6
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
contourpy==1.0.6
cryptoassets.core==0.2
cryptoassets.django==0.1
cryptography==38.0.1
cycler==0.11.0
Django==4.1.2
django-cors-headers==3.13.0
django-cryptapi==0.3.1
django-ninja==0.18.0
django-ninja-auth==0.1.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
ecdsa==0.18.0
fonttools==4.38.0
greenlet==1.1.3.post0
idna==3.4
kiwisolver==1.4.4
logutils==0.3.5
matplotlib==3.6.1
mpmath==1.2.1
mysql-client==0.0.1
mysql-connector==2.2.9
ninja==1.11.1
numpy==1.23.4
packaging==21.3
Pillow==9.3.0
pycodestyle==2.9.1
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.8.2
PyJWT==2.6.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-bitcoinrpc==1.0
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
rainbow-logging-handler==2.2.2
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.42
sqlparse==0.4.3
sympy==1.11.1
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
